Massively improved analysis performance/accuracy by using an "HSB cube"
The original image is sorted into a cube, so the new image knows exactly where to start looking for matches. Sometimes this takes longer, if very few of the pixels match. In this case, it swaps to the legacy analysis mode, which is now stored in a separate file. The UI is also split into its own file.