More shader work, improvements to the coordsImage
The coordsImage wasn't storing enough data, since coordinates need to be from 0-width/height, and they were only from 0-255 for x and y. Now it stores the data using all 24 color bits, or 0-4096 for x and y. It can also be visualized with the UI.

In addition, moved various files into a temp folder to keep them from producing errors while I work on this.