RGB cube samples are now limited capacity
If the RGB cube analysis tries to add a new index to a full arraylist, it will simply replace a random element. This should prevent scenarios where the program slows to a crawl because the start image has a very limited color palette, and so all of the integers go into very few of the arraylists.